Add target_session_attrs support for PG 14+ servers
This adds supports for target_session_attrs, allowing pgbouncer to only
connect to a database's host if it is in a given state, such as primary
or standby.

The implementation relies on specific params from the server that are
only passed starting with PG 14, and therefore it only supports PG 14 or
later. On older versions, target_session_attrs has no effect.

(Small test repair because the host_strategy tests expected 127.0.0.2 to
fail, whereas this adds a replica on that local address.)