Move target_session_attrs check after declaring server ready
Wait until we establish a server is healthy before exiting due to an
invalid session state. This way, we don't consider a server "failed" and
therefore we don't wait to try the next server in the list.

This does mean we might spin through a list of servers if none are in an
expected session state until we reach client_login_timeout.